Add --test option to update-doc.sh
Passing this will skip cloning the repository each time. Useful for
local testing during website development.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk.github.io/+/7242
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ update-doc.sh
@@ -1,12 +1,32 @@
 #!/usr/bin/env bash
 
+set -x
 set -e
 
 cd $(dirname $0)
-
 repo=$(dirname $0)/spdk
 
-git clone --depth 1 https://github.com/spdk/spdk $repo
+TEST_MODE=0
+
+for i in "$@"; do
+        case "$i" in
+                --test)
+                        TEST_MODE=1
+                        ;;
+        esac
+done
+
+if [ -d "$repo" ]; then
+        if [ "$TEST_MODE" -eq 1 ]; then
+                (cd $repo; git clean -x -d -f -f; git checkout master)
+        else
+                rm -rf $repo
+        fi
+fi
+
+if [ ! -d "$repo" ]; then
+        git clone --depth 1 https://github.com/spdk/spdk $repo
+fi
 
 doc_version=$(cd $repo; git rev-parse HEAD)
 
@@ -14,13 +34,16 @@ doc_version=$(cd $repo; git rev-parse HEAD)
 # Overwrite header and footer with the spdk.io versions
 cp _doc_header.html $repo/doc/header.html
 cp _doc_footer.html $repo/doc/footer.html
-cp _doc_stylesheet.css $repo/doc/stylesheet.css
+cp css/spdk.css $repo/doc/spdk.css
 
 (cd $repo/doc; make clean; make)
-git rm -rf doc
+
+rm -rf doc
 cp -R $repo/doc/output/html doc
-git add doc
-rm -rf $repo
+if [ "$TEST_MODE" -eq 0 ]; then
+        git add doc
+        rm -rf $repo
+fi
 
 echo "$doc_version" > _doc_version.txt
 git add _doc_version.txt
